requirements.txt
gemma>=0.1.0
jax>=0.4.20
jaxlib>=0.4.20
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
flask>=2.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
requests>=2.28.0
